move common folder, remove unused import
