--- components/twitch_developer_app/app/common/types.ts
@@ -1,3 +1,0 @@
-export interface Condition {
-  broadcaster_user_id?: string,
-}
