+++ components/twitch_developer_app/common/types.ts
@@ -1,0 +1,3 @@
+export interface Condition {
+  broadcaster_user_id?: string;
+}
