+++ components/twitch_developer_app/sources/streams-by-streamer/streams-by-streamer.ts
@@ -1,6 +1,5 @@
-import common from "../../app/common/common";
+import common from "../../common/common";
 import { defineSource } from "@pipedream/types";
-import { Condition } from "../../app/common/types";
 
 export default defineSource({
   ...common,
